+++ plugins/builtin/programs/src/molecular_visualizer/control_elements/appearance.py
@@ -1,7 +1,7 @@
 from typing import TYPE_CHECKING
 
 from PySide6.QtGui import QColor
-from PySide6.QtWidgets import QHBoxLayout, QLabel, QVBoxLayout
+from PySide6.QtWidgets import QComboBox, QHBoxLayout, QLabel, QVBoxLayout
 
 from mir_commander.core.graphics.utils import color4f_to_qcolor, qcolor_to_color4f
 from mir_commander.ui.sdk.widget import ColorButton
@@ -27,13 +27,29 @@ class Appearance(ControlBlock):
         self._bg_color_button.color_changed.connect(self._bg_color_button_color_changed_handler)
         bg_color_layout.addWidget(self._bg_color_button)
         bg_color_layout.addStretch(1)
-
         layout.addLayout(bg_color_layout)
+
+        style_layout = QHBoxLayout()
+        style_layout.addWidget(QLabel(self.tr("Style:"), self))
+        self._style_combobox = QComboBox()
+        self._style_combobox.currentTextChanged.connect(self._style_combobox_current_text_changed_handler)
+        style_layout.addWidget(self._style_combobox)
+        style_layout.addStretch(1)
+        layout.addLayout(style_layout)
+
         self.setLayout(layout)
 
     def _bg_color_button_color_changed_handler(self, color: QColor):
         self._control_panel.program_action_signal.emit("appearance.set_bg_color", {"color": qcolor_to_color4f(color)})
 
+    def _style_combobox_current_text_changed_handler(self, text: str):
+        self._control_panel.program_action_signal.emit("appearance.set_style", {"name": text})
+
     def update_values(self, program: "Program"):
         color = *program.visualizer.background_color[:3], 1.0
         self._bg_color_button.set_color(color4f_to_qcolor(color))
+
+        for style in program.config.styles:
+            if self._style_combobox.findText(style.name) < 0:
+                self._style_combobox.addItem(style.name)
+        self._style_combobox.setCurrentText(program.visualizer.current_style.name)