+++ CHANGELOG.md
@@ -19,7 +19,7 @@
 #### Molecular visualizer
 
 - Added "background color" option to Save Image dialog
-- Added Appearance control block with "background color" option
+- Added Appearance control block with "background color" option and style
 
 ## v0.1.1 – UI improvements and bug fixes